make work on browser (almost)
