+++ src/lwjgl/java/org/bouncycastle/asn1/ASN1Encodable.java
@@ -1,0 +1,6 @@
+package org.bouncycastle.asn1;
+
+public interface ASN1Encodable
+{
+    ASN1Primitive toASN1Primitive();
+}
