+++ src/lwjgl/java/org/bouncycastle/asn1/ASN1ObjectIdentifier.java
@@ -1,0 +1,14 @@
+package org.bouncycastle.asn1;
+
+public class ASN1ObjectIdentifier extends DERObjectIdentifier
+{
+    public ASN1ObjectIdentifier(String par1Str)
+    {
+        super(par1Str);
+    }
+
+    public ASN1ObjectIdentifier branch(String par1Str)
+    {
+        return new ASN1ObjectIdentifier(this.getId() + "." + par1Str);
+    }
+}
