+++ src/lwjgl/java/org/bouncycastle/jcajce/provider/config/ConfigurableProvider.java
@@ -1,0 +1,5 @@
+package org.bouncycastle.jcajce.provider.config;
+
+public interface ConfigurableProvider
+{
+}
