+++ src/lwjgl/java/org/bouncycastle/jcajce/provider/config/ProviderConfiguration.java
@@ -1,0 +1,5 @@
+package org.bouncycastle.jcajce.provider.config;
+
+public interface ProviderConfiguration
+{
+}
