+++ src/lwjgl/java/org/bouncycastle/jcajce/provider/util/AlgorithmProvider.java
@@ -1,0 +1,8 @@
+package org.bouncycastle.jcajce.provider.util;
+
+import org.bouncycastle.jcajce.provider.config.ConfigurableProvider;
+
+public abstract class AlgorithmProvider
+{
+    public abstract void configure(ConfigurableProvider var1);
+}
