+++ src/main/java/net/minecraft/src/Packet253ServerAuthData.java
@@ -1,5 +1,7 @@
 package net.minecraft.src;
 
+import me.ayunami2000.ayuncraft.CryptManager;
+
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
