--- src/main/java/org/bouncycastle/asn1/ASN1Encodable.java
@@ -1,6 +1,0 @@
-package org.bouncycastle.asn1;
-
-public interface ASN1Encodable
-{
-    ASN1Primitive toASN1Primitive();
-}
