--- src/main/java/org/bouncycastle/jcajce/provider/config/ConfigurableProvider.java
@@ -1,5 +1,0 @@
-package org.bouncycastle.jcajce.provider.config;
-
-public interface ConfigurableProvider
-{
-}
