--- src/main/java/org/bouncycastle/jcajce/provider/config/ProviderConfiguration.java
@@ -1,5 +1,0 @@
-package org.bouncycastle.jcajce.provider.config;
-
-public interface ProviderConfiguration
-{
-}
