--- src/main/java/org/bouncycastle/jcajce/provider/util/AlgorithmProvider.java
@@ -1,8 +1,0 @@
-package org.bouncycastle.jcajce.provider.util;
-
-import org.bouncycastle.jcajce.provider.config.ConfigurableProvider;
-
-public abstract class AlgorithmProvider
-{
-    public abstract void configure(ConfigurableProvider var1);
-}
