+++ src/teavm/java/java/security/Key.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 1996, 2013, Oracle and/or its affiliates. All rights reserved.
+ * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ */
+
+package java.security;
+
+public interface Key {
+    public String getAlgorithm();
+
+    public String getFormat();
+
+    public byte[] getEncoded();
+}
