+++ src/teavm/java/javax/security/auth/Destroyable.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 1999, 2013, Oracle and/or its affiliates. All rights reserved.
+ * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ */
+
+package javax.security.auth;
+
+public interface Destroyable {
+    public default void destroy() throws DestroyFailedException {
+        throw new DestroyFailedException();
+    }
+
+    public default boolean isDestroyed() {
+        return false;
+    }
+}
